+++ src/main/java/com/projectsassy/sassy/userItem/domain/UserItem.java
@@ -1,6 +1,7 @@
-package com.projectsassy.sassy.user.domain;
+package com.projectsassy.sassy.userItem.domain;
 
 import com.projectsassy.sassy.item.domain.Item;
+import com.projectsassy.sassy.user.domain.User;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
