+++ src/main/java/com/projectsassy/sassy/userItem/service/UserItemService.java
@@ -1,15 +1,13 @@
-package com.projectsassy.sassy.user.service;
+package com.projectsassy.sassy.userItem.service;
 
 import com.projectsassy.sassy.common.code.ErrorCode;
 import com.projectsassy.sassy.common.exception.CustomIllegalStateException;
 import com.projectsassy.sassy.item.domain.Badge;
-import com.projectsassy.sassy.item.domain.Item;
-import com.projectsassy.sassy.item.service.ItemService;
 import com.projectsassy.sassy.user.domain.User;
-import com.projectsassy.sassy.user.domain.UserItem;
+import com.projectsassy.sassy.userItem.domain.UserItem;
 import com.projectsassy.sassy.user.dto.UserBadgeDto;
-import com.projectsassy.sassy.user.dto.userItem.AttachBadgeRequest;
-import com.projectsassy.sassy.user.repository.UserItemRepository;
+import com.projectsassy.sassy.userItem.dto.AttachBadgeRequest;
+import com.projectsassy.sassy.userItem.repository.UserItemRepository;
 import com.projectsassy.sassy.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -25,8 +23,6 @@ public class UserItemService {
 
     private final UserRepository userRepository;
     private final UserItemRepository userItemRepository;
-    private final UserService userService;
-    private final ItemService itemService;
 
     public List<UserBadgeDto> findBadges(Long userId) {
         User user = userRepository.findItemsById(userId)
@@ -40,7 +36,7 @@ public class UserItemService {
     }
 
     @Transactional
-    public void attachBadge(Long userId, AttachBadgeRequest attachBadgeRequest) {
+    public void changeBadgeImage(Long userId, AttachBadgeRequest attachBadgeRequest) {
         UserItem userItem = userItemRepository.findUserItem(userId, attachBadgeRequest.getItemId())
             .orElseThrow(() -> {
                 throw new CustomIllegalStateException(ErrorCode.NOT_FOUND_USERITEM);
